Brings in 26 upstream commits (through c2237de04), including the new
FastInterp interpreter, nearest-neighbor scaling in BlitScale/BlitTextures,
Vulkan null descriptors and zero-area renderpass skips, streamed HTTP body
downloads, and the 64-bit config integer promotion.

Conflict resolutions:
- README.md: keep the fork's public-facing overview.
- renderer_base.h: take upstream's u64 current_frame, keep the Android
  eco-turbo budget fields.
- vk_texture_runtime.cpp: keep the fork's RecordParams refactor of
  BlitScale, with the filter set to vk::Filter::eNearest per upstream.

+++ src/video_core/renderer_base.h
@@ -1,4 +1,4 @@
-// Copyright 2014-2025 Citra Emulator Project / Azahar Emulator Project
+// Copyright 2014-2026 Citra Emulator Project / Azahar Emulator Project
 // Licensed under GPLv2 or any later version
 // Refer to the license.txt file included.
 
@@ -117,7 +117,7 @@ public:
         return current_fps;
     }
 
-    s32 GetCurrentFrame() const {
+    u64 GetCurrentFrame() const {
         return current_frame;
     }
 
@@ -156,7 +156,7 @@ protected:
 
 protected:
     f32 current_fps = 0.0f; /// Current framerate, should be set by the renderer
-    s32 current_frame = 0;  /// Current frame, should be set by the renderer
+    u64 current_frame = 0;  /// Current frame, should be set by the renderer
 #ifdef ANDROID
     std::chrono::steady_clock::time_point eco_turbo_budget_update =
         std::chrono::steady_clock::now();
